+++ examples/app/src/main.ts
@@ -1,21 +1,11 @@
-import { Container, Environment, IEnvironmentVariables } from "container.ts";
+import { Container, Environment } from "container.ts";
 import { Assets, Process, Scripts } from "container.ts/lib/node/modules";
 import * as process from "process";
 import { argv } from "yargs";
 import { ERollbarLogsEnv, Main, RollbarLogs, StatsdMetrics, WinstonLogs } from "./modules";
 
-// Environment variables provided by FuseBox build.
-// Fallback to empty if not defined.
-declare const __process_env__: IEnvironmentVariables;
-let fuseBoxEnvironment: IEnvironmentVariables;
-try {
-  fuseBoxEnvironment = __process_env__;
-} catch (error) {
-  fuseBoxEnvironment = {};
-}
-
 // Create environment instance using process environment.
-const ENVIRONMENT = new Environment(fuseBoxEnvironment, process.env);
+const ENVIRONMENT = new Environment(process.env);
 
 // Create container instance with name and environment.
 // Populate container for dependency injection.
@@ -29,18 +19,17 @@ const CONTAINER = new Container("Main", ENVIRONMENT, argv).registerModules([
 ]);
 
 // Register additional modules based on environment definitions.
-if (!!ENVIRONMENT.get(ERollbarLogsEnv.AccessToken)) {
+if (ENVIRONMENT.has(ERollbarLogsEnv.AccessToken)) {
   CONTAINER.registerModule(RollbarLogs);
 }
 
 // Signal operational.
-CONTAINER.up().subscribe({
-  next: () => {
+CONTAINER.up()
+  .then((count) => {
     // Modules counter.
-    CONTAINER.debug(`${CONTAINER.moduleNames.length} modules started`);
-  },
-  error: (error) => {
+    CONTAINER.debug(`${count} modules started`);
+  })
+  .catch((error) => {
     process.stderr.write(`${error}\n`);
     process.exit(1);
-  }
-});
+  });