+++ examples/app/src/modules/main.ts
@@ -1,4 +1,4 @@
-import { IModuleDependencies, Module } from "container.ts";
+import { IModuleDependencies, IModuleHook, Module } from "container.ts";
 import { Scripts, ScriptsProcess } from "container.ts/lib/node/modules";
 import { Observable } from "rxjs";
 
@@ -10,16 +10,18 @@ export class Main extends Module {
   public readonly workerName = "Worker";
   public worker$?: Observable<ScriptsProcess>;
 
-  public moduleDependencies(...prev: IModuleDependencies[]): IModuleDependencies {
+  public moduleDependencies(...prev: IModuleDependencies[]) {
     return super.moduleDependencies(...prev, { scripts: Scripts });
   }
 
-  public moduleUp(): void {
-    // Start worker process.
-    this.worker$ = this.scripts.startWorker(this.workerName, "worker.js");
+  public moduleUp(...args: IModuleHook[]) {
+    return super.moduleUp(...args, async () => {
+      // Start worker process.
+      this.worker$ = this.scripts.startWorker(this.workerName, "worker.js");
 
-    this.worker$.subscribe(() => {
-      this.debug("worker started");
+      this.worker$.subscribe(() => {
+        this.debug("worker started");
+      });
     });
   }
 }